Conflicts:
	project/jni/application/vcmi/AndroidAppSettings.cfg
	project/jni/application/vcmi/Makefile

+++ project/jni/application/openarena/oa-reupload-vm.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 adb shell rm /sdcard/Android/data/ws.openarena.sdl/files/libsdl-DownloadFinished-10.flag
-adb shell rm -r /sdcard/Android/data/ws.openarena.sdl/files/.openarena
+[ -n "$1" ] && adb shell rm -r /sdcard/Android/data/ws.openarena.sdl/files/.openarena
